+++ wp-content/plugins/core/src/Service_Providers/REST_API_Provider.php
@@ -11,23 +11,26 @@ namespace Tribe\Project\Service_Providers;
 
 use Pimple\Container;
 use Tribe\Project\Container\Service_Provider;
-use Tribe\Project\REST_API\REST_API_Main;
+use Tribe\Project\REST_API\Fields\Example_One_Field;
 
 class REST_API_Provider extends Service_Provider {
-	const MAIN = 'rest.main';
+	const EXAMPLE_FIELD = 'rest.fields.example';
 
 	public function register( Container $container ) {
-		$container[ self::MAIN ] = function (): REST_API_Main {
-			return new REST_API_Main();
+		$this->fields( $container );
+	}
+
+	private function fields( Container $container ) {
+		$this->example_field( $container );
+	}
+
+	private function example_field( Container $container ) {
+		$container[ self::EXAMPLE_FIELD ] = function (): Example_One_Field {
+			return new Example_One_Field();
 		};
 
-		add_action(
-			'rest_api_init',
-			function () use ( $container ) {
-				$container['rest.main']->hook();
-			},
-			0,
-			0
-		);
+		add_action( 'rest_api_init', function () use ( $container ) {
+			$container[ self::EXAMPLE_FIELD ]->register_field();
+		}, 0, 0 );
 	}
 }